Relocated exam project
