+++ EXAM_HW/src/main/java/com/example/SpringApp.java
@@ -1,4 +1,4 @@
-package com.example.springapp;
+package com.example;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
