+++ EXAM_HW/src/main/java/com/example/dao/ExamDao.java
@@ -1,10 +1,9 @@
-package com.example.springapp.dao;
+package com.example.dao;
 
-import com.example.springapp.domain.Exam;
+import com.example.domain.Exam;
 import org.springframework.context.MessageSource;
 import org.springframework.core.io.ClassPathResource;
 import java.util.List;
-import java.util.Scanner;
 
 public interface ExamDao {
